requirements.txt
asgiref
Django
sqlparse
psycopg2
gunicorn
Pillow

